requirements.txt
Flask
gunicorn
numpy
tensorflow-cpu==2.12.0
Pillow
gdown
